Removed unnecessary variables. For example, data[1] was the randomly generated value that was stored in another variable called randomBytes. This is completely unnecessary.
All the arrays are the exact same size so there is no reason to keep
re-allocating the array when it just be done in memory. For example, if
I store the encrypted file in data[0], I can just update data[0] with
the random value from data[1] then I can convert data[0] to its
encrypted form.